8-26/flask-test/requirements.txt
Flask==0.10.1
Flask_RESTful==0.3.5
Flask_SQLAlchemy
Flask_Excel==0.0.5
python_memcached==1.58
